% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R
\docType{class}
\name{BindingSitesManager}
\alias{BindingSitesManager}
\alias{.BindingSitesManager}
\title{import shiny}
\usage{
BindingSitesManager(organism, genome, quiet = TRUE)
}
\arguments{
\item{organism}{A character string, one of the supported species names:  hsapiens, mmuscuulus}

\item{genome}{A character string, one of the supported genome builds: hg38, mm10}

\item{quiet}{A logical indicating whether or not the Trena object should print output}
}
\value{
An object of the BindingSitesManager class
}
\description{
a shiny app
}
