% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaViz.R
\docType{methods}
\name{createApp,TrenaViz-method}
\alias{createApp,TrenaViz-method}
\alias{createApp}
\title{Create a runnable shiny app}
\usage{
\S4method{createApp}{TrenaViz}(obj)
}
\arguments{
\item{obj}{An object of class TrenaViz}
}
\description{
Create a runnable shiny app
}
