% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R
\docType{methods}
\name{createDialog,BindingSitesManager-method}
\alias{createDialog,BindingSitesManager-method}
\alias{createDialog}
\title{create and return the control-rich UI}
\usage{
\S4method{createDialog}{BindingSitesManager}(obj)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}
}
\description{
create and return the control-rich UI
}
