% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifsInRegionsBuilder.R
\name{createMotifModelBuilderPage,MotifsInRegionsBuilder-method}
\alias{createMotifModelBuilderPage,MotifsInRegionsBuilder-method}
\alias{createMotifModelBuilderPage}
\title{create and return the control-rich UI}
\usage{
\S4method{createMotifModelBuilderPage}{MotifsInRegionsBuilder}(obj)
}
\arguments{
\item{obj}{An object of class MotifsInRegionsBuilder}
}
\description{
create and return the control-rich UI
}
