% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R,
%   R/FimoDatabaseModelBuilder.R, R/GeneGeneExpressionPlotter.R,
%   R/NetworkView.R, R/TrenaVizModuleExample.R
\docType{methods}
\name{createPage,BindingSitesManager-method}
\alias{createPage,BindingSitesManager-method}
\alias{createPage}
\alias{createPage,FimoDatabaseModelBuilder-method}
\alias{createPage,GeneGeneExpressionPlotter-method}
\alias{createPage,NetworkView-method}
\alias{createPage,ModuleExample-method}
\title{create and return the control-rich UI}
\usage{
\S4method{createPage}{BindingSitesManager}(obj)

\S4method{createPage}{FimoDatabaseModelBuilder}(obj)

\S4method{createPage}{GeneGeneExpressionPlotter}(obj)

\S4method{createPage}{NetworkView}(obj)

\S4method{createPage}{ModuleExample}(obj)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}

\item{obj}{An object of class FimoDatabaseModelBuilder}

\item{obj}{An object of class GeneGeneExpressionPlotter}

\item{obj}{An object of class NetworkView}

\item{obj}{An object of class ModuleExample}
}
\description{
create and return the control-rich UI

create and return the control-rich UI

create and return the control-rich UI

create and return the control-rich UI

create and return the control-rich UI
}
