% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkView.R
\docType{methods}
\name{cyjLayout,NetworkView-method}
\alias{cyjLayout,NetworkView-method}
\alias{cyjLayout}
\title{provide html layout to the caller}
\usage{
\S4method{cyjLayout}{NetworkView}(obj)
}
\arguments{
\item{obj}{An object of class NetworkView}
}
\description{
provide html layout to the caller
}
