% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R,
%   R/FimoModelWidget.R, R/GeneGeneExpressionPlotter.R, R/ModuleExample.R,
%   R/MotifsInRegionsBuilder.R, R/NetworkView.R
\name{displayPage,BindingSitesManager-method}
\alias{displayPage,BindingSitesManager-method}
\alias{displayPage}
\alias{displayPage,FimoModelWidget-method}
\alias{displayPage,GeneGeneExpressionPlotter-method}
\alias{displayPage,ModuleExample-method}
\alias{displayPage,MotifsInRegionsBuilder-method}
\alias{displayPage,NetworkView-method}
\title{display the page}
\usage{
\S4method{displayPage}{BindingSitesManager}(obj, tf)

\S4method{displayPage}{FimoModelWidget}(obj)

\S4method{displayPage}{GeneGeneExpressionPlotter}(obj, tf)

\S4method{displayPage}{ModuleExample}(obj)

\S4method{displayPage}{MotifsInRegionsBuilder}(obj, tf)

\S4method{displayPage}{NetworkView}(obj)
}
\arguments{
\item{obj}{An object of class NetworkView}

\item{tf}{character string, the geneSymbol name of the transcription factor}
}
\description{
display the page

display the page

display the page

display the page

display the page

display the page
}
