% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimoDatabaseModelBuilder.R
\docType{class}
\name{FimoDatabaseModelBuilder}
\alias{FimoDatabaseModelBuilder}
\alias{.FimoDatabaseModelBuilder}
\title{import shiny}
\usage{
FimoDatabaseModelBuilder(message, quiet = TRUE)
}
\arguments{
\item{message}{A character string, one of the supported species names:  hsapiens, mmuscuulus}
}
\value{
An object of the FimoDatabaseModelBuilder class
}
\description{
a shiny app
}
