% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneGeneExpressionPlotter.R
\docType{class}
\name{GeneGeneExpressionPlotter}
\alias{GeneGeneExpressionPlotter}
\alias{.GeneGeneExpressionPlotter}
\title{import shiny}
\usage{
GeneGeneExpressionPlotter(quiet = TRUE)
}
\arguments{
\item{quiet}{A logical indicating whether or not the Trena object should print output}
}
\value{
An object of the GeneGeneExpressionPlotter class
}
\description{
a shiny app
}
