% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMethodMotifMatcher.R
\docType{methods}
\name{match,MultiMethodMotifMatcher-method}
\alias{match,MultiMethodMotifMatcher-method}
\title{run the MultiMethodMotifMatcher}
\usage{
\S4method{match}{MultiMethodMotifMatcher}(obj)
}
\arguments{
\item{obj}{A properly configured MultiMethodMotifMatching object}
}
\value{
A data.frame with scored hits
}
\description{
run the MultiMethodMotifMatcher
}
