% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleExample.R
\docType{class}
\name{ModuleExample}
\alias{ModuleExample}
\alias{.ModuleExample}
\title{import shiny}
\usage{
ModuleExample(quiet = TRUE)
}
\arguments{
\item{message}{A character string, one of the supported species names:  hsapiens, mmuscuulus}
}
\value{
An object of the ModuleExample class
}
\description{
a shiny app
}
