% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifsInRegionsBuilder.R
\docType{class}
\name{MotifsInRegionsBuilder}
\alias{MotifsInRegionsBuilder}
\alias{.MotifsInRegionsBuilder}
\title{import shiny}
\usage{
MotifsInRegionsBuilder(quiet = TRUE)
}
\arguments{
\item{quiet}{A logical indicating whether or not the Trena object should print output}

\item{organism}{A character string, one of the supported species names:  hsapiens, mmuscuulus}

\item{genome}{A character string, one of the supported genome builds: hg38, mm10}
}
\value{
An object of the MotifsInRegionsBuilder class
}
\description{
a shiny app
}
