% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMethodMotifMatcher.R
\docType{class}
\name{MultiMethodMotifMatcher}
\alias{MultiMethodMotifMatcher}
\alias{.MultiMethodMotifMatcher}
\title{import trena
import motifmatchr
import universalmotif
import TFBSTools}
\usage{
MultiMethodMotifMatcher(genome, motifMatrix, tbl.regions,
  sequenceMatchAlgorithm, matchThreshold, quiet = TRUE)
}
\arguments{
\item{genome}{A character string, one of the supported genome builds: hg38, mm10}

\item{motifMatrix}{A position frequency matrix}

\item{sequenceMatchAlgorithm}{A character string, either "Biostrings matchPWM" or "MOODS matchMotifs"}

\item{matchThreshold}{numeric, between 0 and 1 (usually 0.7 or higher)}

\item{quiet}{A logical indicating whether or not the Trena object should print output}

\item{regions}{A data.frame with chrom, start, end columns}
}
\value{
An object of the MultiMethodMotifMatcher class
}
\description{
provides unified access to a set of motif/DNA sequence matching algorithms
}
