% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkView.R
\docType{class}
\name{NetworkView}
\alias{NetworkView}
\alias{.NetworkView}
\title{import shiny
import cyjShiny
import TrenaProject
import graph}
\usage{
NetworkView(targetGene, tbl.model, tbl.regulatoryRegions, expressionMatrix,
  quiet = TRUE)
}
\arguments{
\item{quiet}{A logical indicating whether or not the Trena object should print output}

\item{organism}{A character string, one of the supported species names:  hsapiens, mmuscuulus}

\item{genome}{A character string, one of the supported genome builds: hg38, mm10}
}
\value{
An object of the NetworkView class
}
\description{
a shiny app
}
