% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R
\name{removeLogos,BindingSitesManager-method}
\alias{removeLogos,BindingSitesManager-method}
\alias{removeLogos}
\title{remove the image}
\usage{
\S4method{removeLogos}{BindingSitesManager}(obj)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}
}
\description{
remove the image
}
