% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R
\docType{methods}
\name{renderLogos,BindingSitesManager-method}
\alias{renderLogos,BindingSitesManager-method}
\alias{renderLogos}
\title{show the control-rich UI}
\usage{
\S4method{renderLogos}{BindingSitesManager}(obj, tfMappingOption)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}
}
\description{
show the control-rich UI
}
