% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneGeneExpressionPlotter.R
\docType{methods}
\name{setGenes,GeneGeneExpressionPlotter-method}
\alias{setGenes,GeneGeneExpressionPlotter-method}
\alias{setGenes}
\title{specify the pair of genes, typically a TF and a targetGene}
\usage{
\S4method{setGenes}{GeneGeneExpressionPlotter}(obj, gene1, gene2)
}
\arguments{
\item{obj}{An object of class GeneGeneExpressionPlotter}

\item{gene1}{A character string}

\item{gene2}{A character string}
}
\description{
specify the pair of genes, typically a TF and a targetGene
}
