% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R
\docType{methods}
\name{setGenome,BindingSitesManager-method}
\alias{setGenome,BindingSitesManager-method}
\alias{setGenome}
\title{specity the genome build: hg38, mm10, ...}
\usage{
\S4method{setGenome}{BindingSitesManager}(obj, genomeName)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}

\item{genome}{build character string: hg38, mm10, ...}
}
\description{
specity the genome build: hg38, mm10, ...
}
