% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R, R/FimoModelWidget.R,
%   R/ModuleExample.R
\docType{methods}
\name{setGenomicRegion,BindingSitesManager-method}
\alias{setGenomicRegion,BindingSitesManager-method}
\alias{setGenomicRegion}
\alias{setGenomicRegion,FimoModelWidget-method}
\alias{setGenomicRegion,ModuleExample-method}
\alias{setGenomciRegion}
\title{specify the one or more genomic regions in which to look for binding sites}
\usage{
\S4method{setGenomicRegion}{BindingSitesManager}(obj, tbl.region)

\S4method{setGenomicRegion}{FimoModelWidget}(obj, tbl.region)

\S4method{setGenomicRegion}{ModuleExample}(obj, tbl.region)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}

\item{tbl.region}{a data.frame containing a single (?) genonmic region}

\item{obj}{An object of class FimoModelWidget}

\item{tbl.region}{a data.frame of 1 row: chrom, start and end columns}

\item{obj}{An object of class FimoModelWidget}

\item{newRegion}{a character string}
}
\description{
specify the one or more genomic regions in which to look for binding sites

specify the chromosomal location of interest

set the chromosomal locus of current interestg
}
