% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneGeneExpressionPlotter.R
\docType{methods}
\name{setMatrixForPlotting,GeneGeneExpressionPlotter-method}
\alias{setMatrixForPlotting,GeneGeneExpressionPlotter-method}
\alias{setMatrixForPlotting}
\title{specity the expression matrix}
\usage{
\S4method{setMatrixForPlotting}{GeneGeneExpressionPlotter}(obj, matrix)
}
\arguments{
\item{obj}{An object of class GeneGeneExpressionPlotter}

\item{matrix}{the expression matrix}
}
\description{
specity the expression matrix
}
