% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaVizModuleExample.R
\docType{methods}
\name{setMessage,ModuleExample-method}
\alias{setMessage,ModuleExample-method}
\title{assign a new message to be display}
\usage{
\S4method{setMessage}{ModuleExample}(obj, newValue)
}
\arguments{
\item{newValue}{A character string}
}
\description{
assign a new message to be display
}
