% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimoDatabaseModelBuilder.R
\docType{methods}
\name{setMessageX,FimoDatabaseModelBuilder-method}
\alias{setMessageX,FimoDatabaseModelBuilder-method}
\title{assign a new message to be display}
\usage{
\S4method{setMessageX}{FimoDatabaseModelBuilder}(obj, newValue)
}
\arguments{
\item{newValue}{A character string}
}
\description{
assign a new message to be display
}
