% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R
\name{setOrganism,BindingSitesManager-method}
\alias{setOrganism,BindingSitesManager-method}
\alias{setOrganism}
\title{specity the organism: Hsapiens, Mmusculus, ...}
\usage{
\S4method{setOrganism}{BindingSitesManager}(obj, organism)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}

\item{organism}{character string: Hsapiens, Mmusculus}
}
\description{
specity the organism: Hsapiens, Mmusculus, ...
}
