% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimoModelWidget.R
\name{setRegulatoryRegions,FimoModelWidget-method}
\alias{setRegulatoryRegions,FimoModelWidget-method}
\alias{setRegulatoryRegions}
\title{areas of (currently) open chromatin}
\usage{
\S4method{setRegulatoryRegions}{FimoModelWidget}(obj, tbls.regulatoryRegions)
}
\arguments{
\item{obj}{An object of class FimoModelWidget}

\item{tbsl.regulatoryRegions}{a list of data.frames, each with chrom, start and end columns}
}
\description{
areas of (currently) open chromatin
}
