% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimoModelWidget.R
\docType{methods}
\name{setTargetGene,FimoModelWidget-method}
\alias{setTargetGene,FimoModelWidget-method}
\alias{setTargetGene}
\title{supply the name of the regulated gene}
\usage{
\S4method{setTargetGene}{FimoModelWidget}(obj, targetGene)
}
\arguments{
\item{obj}{An object of class FimoModelWidget}

\item{targetGene}{character, a geneSymbol}
}
\description{
supply the name of the regulated gene
}
