% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R
\name{setTF,BindingSitesManager-method}
\alias{setTF,BindingSitesManager-method}
\alias{setTF}
\title{specity the tf (transcription factor) to work on}
\usage{
\S4method{setTF}{BindingSitesManager}(obj, tf)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}

\item{TF}{character string, the transcription factor we are currently concerned with}
}
\description{
specity the tf (transcription factor) to work on
}
