% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardizeMotifMatchedTable}
\alias{standardizeMotifMatchedTable}
\title{standardize motif-matched regions from footprint databases and directly from fimo}
\usage{
standardizeMotifMatchedTable(tbl)
}
\arguments{
\item{tbl}{data.frame}
}
\value{
data.frame
}
\description{
standardize motif-matched regions from footprint databases and directly from fimo
}
