% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trenaViz.R
\docType{data}
\name{trenaViz}
\alias{trenaViz}
\alias{colors}
\title{import shiny
import shinydashboard
import shinyjs
import later
import igvShiny
import DT
import VariantAnnotation
import trenaSGM
importFrom  RColorBrewer brewer.pal}
\format{An object of class \code{character} of length 8.}
\usage{
colors
}
\description{
import shiny
import shinydashboard
import shinyjs
import later
import igvShiny
import DT
import VariantAnnotation
import trenaSGM
importFrom  RColorBrewer brewer.pal
}
\keyword{datasets}
