% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R,
%   R/FimoDatabaseModelBuilder.R, R/GeneGeneExpressionPlotter.R,
%   R/MotifsInRegionsBuilder.R, R/NetworkView.R, R/TrenaVizModuleExample.R
\docType{methods}
\name{addEventHandlers,BindingSitesManager-method}
\alias{addEventHandlers,BindingSitesManager-method}
\alias{addEventHandlers}
\alias{addEventHandlers,FimoDatabaseModelBuilder-method}
\alias{addEventHandlers,GeneGeneExpressionPlotter-method}
\alias{addEventHandlers,MotifsInRegionsBuilder-method}
\alias{addEventHandlers,NetworkView-method}
\alias{addEventHandlers,ModuleExample-method}
\title{add shiny event handlers}
\usage{
\S4method{addEventHandlers}{BindingSitesManager}(obj, session, input,
  output)

\S4method{addEventHandlers}{FimoDatabaseModelBuilder}(obj, session, input,
  output)

\S4method{addEventHandlers}{GeneGeneExpressionPlotter}(obj, session, input,
  output)

\S4method{addEventHandlers}{MotifsInRegionsBuilder}(obj, session, input,
  output)

\S4method{addEventHandlers}{NetworkView}(obj, session, input, output)

\S4method{addEventHandlers}{ModuleExample}(obj, session, input, output)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}

\item{session}{a Shiny session object}

\item{input}{a Shiny input object}

\item{output}{a Shiny output object}

\item{obj}{An object of class FimoDatabaseModelBuilder}

\item{session}{a Shiny session object}

\item{input}{a Shiny input object}

\item{output}{a Shiny output object}

\item{obj}{An object of class BindingSitesManager}

\item{session}{a Shiny session object}

\item{input}{a Shiny input object}

\item{output}{a Shiny output object}

\item{obj}{An object of class MotifsInRegionsBuilder}

\item{session}{a Shiny session object}

\item{input}{a Shiny input object}

\item{output}{a Shiny output object}

\item{obj}{An object of class NetworkView}

\item{session}{a Shiny session object}

\item{input}{a Shiny input object}

\item{output}{a Shiny output object}

\item{obj}{An object of class ModuleExample}

\item{session}{a Shiny session object}

\item{input}{a Shiny input object}

\item{output}{a Shiny output object}
}
\description{
add shiny event handlers

add shiny event handlers

add shiny event handlers

add shiny event handlers

add shiny event handlers

add shiny event handlers
}
