% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R, R/FimoModelWidget.R,
%   R/GeneGeneExpressionPlotter.R, R/ModuleExample.R, R/NetworkView.R
\name{createPage,BindingSitesManager-method}
\alias{createPage,BindingSitesManager-method}
\alias{createPage}
\alias{.fimoBuilderCreatePage,FimoModelWidget-method}
\alias{createPage,GeneGeneExpressionPlotter-method}
\alias{.moduleExampleCreatePage,ModuleExample-method}
\alias{createPage,NetworkView-method}
\title{create and return the control-rich UI}
\usage{
\S4method{createPage}{BindingSitesManager}(obj)

\S4method{.fimoBuilderCreatePage}{FimoModelWidget}(obj)

\S4method{createPage}{GeneGeneExpressionPlotter}(obj)

\S4method{.moduleExampleCreatePage}{ModuleExample}(obj)

\S4method{createPage}{NetworkView}(obj)
}
\arguments{
\item{obj}{An object of class NetworkView}
}
\description{
create and return the control-rich UI

create and return the control-rich UI

create and return the control-rich UI

create and return the control-rich UI

create and return the control-rich UI
}
