% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R,
%   R/FimoDatabaseModelBuilder.R, R/GeneGeneExpressionPlotter.R,
%   R/ModuleExample.R, R/NetworkView.R
\docType{methods}
\name{createPage,BindingSitesManager-method}
\alias{createPage,BindingSitesManager-method}
\alias{createPage}
\alias{.fimoBuilderCreatePage,FimoDatabaseModelBuilder-method}
\alias{createPage,GeneGeneExpressionPlotter-method}
\alias{.moduleExampleCreatePage,ModuleExample-method}
\alias{createPage,NetworkView-method}
\title{create and return the control-rich UI}
\usage{
\S4method{createPage}{BindingSitesManager}(obj)

\S4method{.fimoBuilderCreatePage}{FimoDatabaseModelBuilder}(obj)

\S4method{createPage}{GeneGeneExpressionPlotter}(obj)

\S4method{.moduleExampleCreatePage}{ModuleExample}(obj)

\S4method{createPage}{NetworkView}(obj)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}

\item{obj}{An object of class FimoDatabaseModelBuilder}

\item{obj}{An object of class GeneGeneExpressionPlotter}

\item{obj}{An object of class ModuleExample}

\item{obj}{An object of class NetworkView}
}
\description{
create and return the control-rich UI

create and return the control-rich UI

create and return the control-rich UI

create and return the control-rich UI

create and return the control-rich UI
}
