% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R
\docType{methods}
\name{displayPage,BindingSitesManager-method}
\alias{displayPage,BindingSitesManager-method}
\alias{displayPage}
\title{display the page}
\usage{
\S4method{displayPage}{BindingSitesManager}(obj, tf)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}

\item{tf}{character string, the geneSymbol name of the transcription factor}
}
\description{
display the page
}
