% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R,
%   R/MotifsInRegionsBuilder.R, R/NetworkView.R
\docType{methods}
\name{displayPage,BindingSitesManager-method}
\alias{displayPage,BindingSitesManager-method}
\alias{displayPage}
\alias{displayPage,MotifsInRegionsBuilder-method}
\alias{displayPage,NetworkView-method}
\title{display the page}
\usage{
\S4method{displayPage}{BindingSitesManager}(obj, tf)

\S4method{displayPage}{MotifsInRegionsBuilder}(obj, tf)

\S4method{displayPage}{NetworkView}(obj)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}

\item{tf}{character string, the geneSymbol name of the transcription factor}

\item{obj}{An object of class MotifsInRegionsBuilder}

\item{tf}{character string, the geneSymbol name of the transcription factor}

\item{obj}{An object of class NetworkView}

\item{tf}{character string, the geneSymbol name of the transcription factor}
}
\description{
display the page

display the page

display the page
}
