% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimoDatabaseModelBuilder.R
\docType{class}
\name{FimoDatabaseModelBuilder}
\alias{FimoDatabaseModelBuilder}
\alias{.FimoDatabaseModelBuilder}
\title{Create an FimoDatabaseModelBuilder object}
\usage{
FimoDatabaseModelBuilder(trenaProject, targetGene, genomicRegion,
  tbls.regulatoryRegions, quiet = TRUE)
}
\arguments{
\item{targetGene}{character standard geneSymbol name, eg, "GATA2"}

\item{genomicRegion}{character, eg "chr3:128,475,725-128,498,247"}

\item{trenaProjectName}{character eg, "TrenaProjectErythropoiesis"}
}
\value{
An object of the FimoDatabaseModelBuilder class
}
\description{
a shiny app
}
