% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimoModelWidget.R
\docType{class}
\name{FimoModelWidget}
\alias{FimoModelWidget}
\alias{.FimoModelWidget}
\title{Create an FimoModelWidget object}
\usage{
FimoModelWidget(quiet = TRUE)
}
\arguments{
\item{genomicRegion}{character, eg "chr3:128,475,725-128,498,247"}
}
\value{
An object of the FimoModelWidget class
}
\description{
a shiny app
}
