% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R, R/FimoModelWidget.R
\docType{methods}
\name{setGenomicRegion,BindingSitesManager-method}
\alias{setGenomicRegion,BindingSitesManager-method}
\alias{setGenomicRegion}
\alias{setGenomicRegion,FimoModelWidget-method}
\title{specify the one or more genomic regions in which to look for binding sites}
\usage{
\S4method{setGenomicRegion}{BindingSitesManager}(obj, tbl.region)

\S4method{setGenomicRegion}{FimoModelWidget}(obj, tbl.region)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}

\item{tbl.region}{a data.frame of 1 row: chrom, start and end columns}

\item{TF}{character string, the transcription factor we are currently concerned with}

\item{obj}{An object of class FimoModelWidget}
}
\description{
specify the one or more genomic regions in which to look for binding sites

specify the chromosomal location of interest
}
