% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindingSitesManager.R
\docType{methods}
\name{setGenomicRegion,BindingSitesManager-method}
\alias{setGenomicRegion,BindingSitesManager-method}
\alias{setGenomicRegion}
\title{specify the one or more genomic regions in which to look for binding sites}
\usage{
\S4method{setGenomicRegion}{BindingSitesManager}(obj, tbl.regions)
}
\arguments{
\item{obj}{An object of class BindingSitesManager}

\item{TF}{character string, the transcription factor we are currently concerned with}
}
\description{
specify the one or more genomic regions in which to look for binding sites
}
