% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FimoModelWidget.R, R/ModuleExample.R
\docType{methods}
\name{setTrenaProject,FimoModelWidget-method}
\alias{setTrenaProject,FimoModelWidget-method}
\alias{setTrenaProject}
\alias{setTrenaProject,ModuleExample-method}
\title{supply tissue-specific data}
\usage{
\S4method{setTrenaProject}{FimoModelWidget}(obj, trenaProject)

\S4method{setTrenaProject}{ModuleExample}(obj, trenaProject)

\S4method{setTrenaProject}{ModuleExample}(obj, trenaProject)
}
\arguments{
\item{obj}{An object of class FimoModelWidget}

\item{trenaProject}{an instance of a concrete subclass of TrenaProject}

\item{obj}{An object of class FimoModelWidget}

\item{trenaProject}{an instance of a concrete subclass of TrenaProject}

\item{obj}{An object of class FimoModelWidget}

\item{trenaProject}{an instance of a concrete subclass of TrenaProject}
}
\description{
supply tissue-specific data

supply tissue-specific data

supply tissue-specific data
}
