% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaViz.R
\docType{class}
\name{TrenaViz}
\alias{TrenaViz}
\alias{.TrenaViz}
\title{import shiny
import shinydashboard
import shinyjs
import later
import igvShiny
import DT
import VariantAnnotation
import trenaSGM
importFrom  RColorBrewer brewer.pal}
\usage{
TrenaViz(projectName, quiet = TRUE)
}
\arguments{
\item{projectName}{A string indicating a TrenaProject subclass}

\item{quiet}{A logical indicating whether or not the Trena object should print output}
}
\value{
An object of the TrenaViz class
}
\description{
a shiny app
}
