% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaViz.R
\docType{data}
\name{TrenaViz}
\alias{TrenaViz}
\alias{tbl.region}
\title{import TrenaProject
import shiny
import shinydashboard
import shinyjs
import later
import yaml
import igvShiny
import DT
import VariantAnnotation
import trenaSGM
importFrom  RColorBrewer brewer.pal
import colourpicker
import later
import MotifDb
import ggpseqlogo}
\format{An object of class \code{data.frame} with 1 rows and 3 columns.}
\usage{
tbl.region

TrenaViz(projectName, quiet = TRUE)
}
\arguments{
\item{projectName}{A string indicating a TrenaProject subclass}

\item{quiet}{A logical indicating whether or not the Trena object should print output}
}
\value{
An object of the TrenaViz class
}
\description{
a shiny app
}
\keyword{datasets}
